chore:rename algorithm
