+++ src/Domain/Algorithm/AlgorithmFactory.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, test } from "vitest";
 import { TokenBucket } from "./TokenBucket.ts";
-import { AlgorithmName, rateLimiterFactory } from "./RateLimiterFactory.ts";
+import { AlgorithmName, algorithmFactory } from "./AlgorithmFactory.ts";
 
 describe("RateLimiter Algorithm Factory", () => {
   test("returns TokenBucket instance when algorithm name is TokenBucket", () => {
     const algorithmName = "TokenBucket";
 
-    const actual = rateLimiterFactory(algorithmName);
+    const actual = algorithmFactory(algorithmName);
 
     expect(actual).toBeInstanceOf(TokenBucket);
   });
@@ -14,6 +14,6 @@ describe("RateLimiter Algorithm Factory", () => {
   test("throws error if the algorithm name is unsupported", () => {
     const algorithmName = "fake-algorithm-name" as AlgorithmName;
 
-    expect(() => rateLimiterFactory(algorithmName)).toThrow("fake-algorithm-name is unsupported algorithm");
+    expect(() => algorithmFactory(algorithmName)).toThrow("fake-algorithm-name is unsupported algorithm");
   });
 });
