+++ src/Domain/Algorithm/AlgorithmFactory.ts
@@ -10,7 +10,7 @@ export type AlgorithmsMap = {
 };
 export type AlgorithmName = keyof AlgorithmsMap;
 
-export function rateLimiterFactory<Name extends AlgorithmName>(algorithmName: Name): AlgorithmsMap[Name]["algorithm"] {
+export function algorithmFactory<Name extends AlgorithmName>(algorithmName: Name): AlgorithmsMap[Name]["algorithm"] {
   if (algorithmName === "TokenBucket") return new TokenBucket() as AlgorithmsMap[Name]["algorithm"];
 
   throw new Error(`${algorithmName} is unsupported algorithm.`);
